Nits in server code
